Starting to re-write most of the code
The original version is not compliant with Android 4.0+ StrictMode.
To fix this, all network access needs to be off the main thread.
When finished, this should also make it more flexable and
error-resistant.

So far: The login activity is the main activity, and handles the login
process before starting the main player activity. This prevents being
stuck on the player without being logged in.